requirements.txt
langchain==0.2.0
langchain_community==0.2.0
langchain_core==0.2.20
langchain_ollama==0.1.0
python-dotenv==1.0.1
streamlit==1.31.1
